1. Refactored fields 2. Deleted directories for each files 3. Moved saga for products into one file 4. Split styles
