+++ src/components/index.js
@@ -5,10 +5,10 @@ import { Orders as OrdersComponent } from './Orders';
 export { Shipping } from './Shipping';
 export { Billing } from './Billing';
 export { Payment } from './Payment';
-
-export { BreadcrumbItem } from './BreadcrumbItem';
-export { ButtonComponent } from './ButtonComponent';
-
+export { AddressFields } from './AddressFields';
+export { ContactFields } from './ContactFields';
+export { Breadcrumbs } from './Breadcrumbs';
+export { Button } from './Button';
 export { FinishOrders } from './FinishOrders';
 
 const mapStateToProps = (state) => ({
