--- src/constants/devices.js
@@ -1,6 +1,0 @@
-export const DEVICES = {
-  small: `(min-width: 576px)`,
-  medium: `(min-width: 768px)`,
-  large: `(min-width: 992px)`,
-  extraLarge: `(min-width: 1200px)`,
-};
