+++ src/styled/Button.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Button from 'react-bootstrap/Button';
-import { COLORS } from './colors';
+import { COLORS } from '../constants/colors';
 
 export const StyledButton = styled(Button)`
   background-color: ${COLORS.orchid};
