+++ src/styled/ContactFields.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const InputSpan = styled.span`
+  margin-left: 20px;
+  font-size: 12px;
+`;
