+++ src/styled/FinishOrders.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const OrderWrap = styled.div`
+  margin-top: 50px;
+  margin-left: 16px;
+`;
