+++ src/styled/Payment.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const SecureInfo = styled.div`
+  margin-left: 16px;
+  margin-top: 15px;
+`;
+
+export const Icon = styled.img`
+  width: 25px;
+  height: 25px;
+`;
+
+export const Span = styled.span`
+  font-size: 18px;
+  padding-left: 10px;
+`;
