+++ src/styled/Shipping.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Row from 'react-bootstrap/Row';
-import { COLORS } from './colors';
+import { COLORS } from '../constants/colors';
 
 export const RowWrap = styled(Row)`
   height: 70vh;
