+++ src/styled/shared/FormGroupInner.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const FormGroupInner = styled.div`
+  display: flex;
+`;
