+++ src/styled/shared/InputsWrap.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const InputsWrap = styled.div`
+  margin-top: 25px;
+`;
