+++ src/styled/shared/Label.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+import { COLORS } from '../../constants/colors';
+
+export const Label = styled.label`
+  font-size: 1.2em;
+  color: ${COLORS.orchid};
+`;
