+++ src/styled/shared/Title.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { COLORS } from '../../constants/colors';
+
+export const Title = styled.h1`
+  font-size: 2em;
+  margin-left: 16px;
+  color: ${COLORS.orchid};
+`;
